initial round of feedback
