+++ ui-admin/src/study/participants/enrolleeView/EnrolleeOverview.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import Api, { EnrolleeRelation, Profile } from 'api/api'
+import Api, { EnrolleeRelation } from 'api/api'
 import { StudyEnvContextT } from '../../StudyEnvironmentRouter'
 import ParticipantNotesView from './ParticipantNotesView'
-import { dateToDefaultString, Enrollee } from '@juniper/ui-core'
+import { dateToDefaultString, Enrollee, Profile } from '@juniper/ui-core'
 import KitRequests from '../KitRequests'
 import { InfoCard, InfoCardBody, InfoCardHeader, InfoCardTitle, InfoCardValue } from 'components/InfoCard'
 import { useLoadingEffect } from 'api/api-utils'
