+++ ui-admin/src/study/participants/survey/EnrolleeSurveyView.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import { RawEnrolleeSurveyView } from './EnrolleeSurveyView'
+import { RawEnrolleeSurveyView } from './SurveyResponseView'
 import {
   mockAnswer,
   mockConfiguredSurvey,
